backend/requirements.txt
django
djangorestframework
fastapi
torch
sentence-transformers
pylev
transformers